requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.10.0
requests>=2.31.0
jsonschema>=4.21.1
streamlit>=1.28.0
